Order queue fix and speedmining tweaks
`Commander.commands` being a hashmap causes orders to be passed to sc2 in hash order (i.e. basically random), causing queued orders to be delivered in the wrong order about 50% of the time.

This prevented implementing the key speedmining boost trick, which is to queue the `smart` command on top of the `move` command.

With this fix in place, and the speedmining boost implemented, the speedmining example increases from a 3% bonus to an 8% bonus (the same speed up that Ares achieves in the Python world).